rename popup-scripts to options-page
