+++ src/options/options.html
@@ -11,7 +11,7 @@
       rel="stylesheet"
     />
     <title>YouTube Rewind & Fast Forward Buttons - Options Page</title>
-    <script type="module" src="popup-script.ts" defer></script>
+    <script type="module" src="options-page.ts" defer></script>
     <script
       type="module"
       src="../web-components/support-links.ts"
